(#1) ProcessUtils: extract process-related methods
